Add initial scraping implementation
This implementation is similar to `llui85/covid19stats-govau`, but has been cleaned up a fair bit and commented. This commit changes the scraping method to use websockets, but does not do anything with the raw data except write it to a file.

Two new directories (`raw-new` and `rawhtml-new`) have been created to store raw data in, as backwards compatibility hasn't been dicussed yet.

Puppeteer and cheerio have been removed, as they are no longer necessary. `node-fetch` (for getting graphIds) and `ws` (for connecting to the QLik API) have been added.

Two new configuration options have been added to `scrape.js`. These are:
`PAGE_URI`  - url for the page containing the graphs
`QLIK_WS_PREFIX` - prefix for the QLik Sense server to connect to

- llui85