Add legacy data creation script
This commit adds a new script, which bridges the old raw data system into the new raw data system. This solution should be the most backwards compatible, as only minor details have changed. Therefore, consumers of this data should not be affected.

Changes to the data are that:
- Numbers are no longer comma-separated
- Decimals are no longer truncated
- Percentages are no longer prefixed with a percentage symbol

This script should be run after `scrape.js`, but before running `generate-csv.js`. It generates files in the `data/raw` directory from the files in the `data/raw-new`.

This commit also adds the first set of new data, for the date `2020-07-06`.

- llui85